+++ sdk/communication/azure-communication-phonenumbers/azure/communication/phonenumbers/_api_versions.py
@@ -11,6 +11,8 @@ class ApiVersion(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     V2022_01_11_PREVIEW2 = "2022-01-11-preview2"
     V2022_12_01 = "2022-12-01"
     V2023_2023_10_01_PREVIEW = "2023-10-01-preview"
+    V2023_05_01_PREVIEW = "2023-05-01-preview"
     V2024_01_PREVIEW1 = "2024-01-31-preview"
+    V2024_03_01 = "2024-03-01"
 
 DEFAULT_VERSION = ApiVersion.V2024_01_PREVIEW1