+++ sdk/communication/azure-communication-phonenumbers/azure/communication/phonenumbers/_generated/models/__init__.py
@@ -10,6 +10,7 @@ from ._models import CommunicationError
 from ._models import CommunicationErrorResponse
 from ._models import OperatorDetails
 from ._models import OperatorInformation
+from ._models import OperatorInformationOptions
 from ._models import OperatorInformationRequest
 from ._models import OperatorInformationResult
 from ._models import PhoneNumberAdministrativeDivision
@@ -33,7 +34,7 @@ from ._enums import PhoneNumberAssignmentType
 from ._enums import PhoneNumberCapabilityType
 from ._enums import PhoneNumberOperationStatus
 from ._enums import PhoneNumberOperationType
-from ._enums import PhoneNumberSource
+from ._enums import PhoneNumberSearchResultError
 from ._enums import PhoneNumberType
 from ._patch import __all__ as _patch_all
 from ._patch import *  # pylint: disable=unused-wildcard-import
@@ -44,6 +45,7 @@ __all__ = [
     "CommunicationErrorResponse",
     "OperatorDetails",
     "OperatorInformation",
+    "OperatorInformationOptions",
     "OperatorInformationRequest",
     "OperatorInformationResult",
     "PhoneNumberAdministrativeDivision",
@@ -67,6 +69,7 @@ __all__ = [
     "PhoneNumberOperationStatus",
     "PhoneNumberOperationType",
     "PhoneNumberSource",
+    "PhoneNumberSearchResultError",
     "PhoneNumberType",
 ]
 __all__.extend([p for p in _patch_all if p not in __all__])