+++ sdk/communication/azure-communication-phonenumbers/test/test_phone_number_administration_client_async.py
@@ -221,7 +221,7 @@ class TestPhoneNumbersClientAsync(PhoneNumbersTestCase):
             )
             phone_number_to_buy = await search_poller.result()
             purchase_poller = await phone_number_client.begin_purchase_phone_numbers(
-                phone_number_to_buy.search_id, consent_dnr=True, polling=True)
+            phone_number_to_buy.search_id, no_resale_consent=True, polling=True)
 
             await purchase_poller.result()
             assert purchase_poller.status() == PhoneNumberOperationStatus.SUCCEEDED.value
@@ -275,7 +275,7 @@ class TestPhoneNumbersClientAsync(PhoneNumbersTestCase):
             )
             phone_number_to_buy = await search_poller.result()
             purchase_poller = await self.phone_number_client.begin_purchase_phone_numbers(
-                phone_number_to_buy.search_id, consent_dnr=True, polling=True)
+                phone_number_to_buy.search_id, no_resale_consent=True, polling=True)
 
             await purchase_poller.result()
             assert purchase_poller.status() == PhoneNumberOperationStatus.SUCCEEDED.value
@@ -517,3 +517,42 @@ class TestPhoneNumbersClientAsync(PhoneNumbersTestCase):
             async for item in offerings:
                 items.append(item)
         assert len(items) > 0
+
+    @recorded_by_proxy_async
+    async def test_search_operator_information_with_too_many_phone_numbers(self):
+        if self.is_playback():
+            phone_numbers = [ "sanitized", "sanitized" ]
+        else:
+            phone_numbers = [ self.phone_number, self.phone_number ]
+
+        with pytest.raises(Exception) as ex:
+            async with self.phone_number_client:
+                await self.phone_number_client.search_operator_information(phone_numbers)
+
+        assert is_client_error_status_code(
+            ex.value.status_code) is True, 'Status code {ex.value.status_code} does not indicate a client error'  # type: ignore
+        assert ex.value.message is not None  # type: ignore
+
+    @recorded_by_proxy_async
+    async def test_search_operator_information_with_list(self):
+        if self.is_playback():
+            phone_number = "sanitized"
+        else:
+            phone_number = self.phone_number
+
+        async with self.phone_number_client:
+            results = await self.phone_number_client.search_operator_information([ phone_number ])
+        assert len(results.values) == 1
+        assert results.values[0].phone_number == self.phone_number
+
+    @recorded_by_proxy_async
+    async def test_search_operator_information_with_single_string(self):
+        if self.is_playback():
+            phone_number = "sanitized"
+        else:
+            phone_number = self.phone_number
+
+        async with self.phone_number_client:
+            results = await self.phone_number_client.search_operator_information(phone_number)
+        assert len(results.values) == 1
+        assert results.values[0].phone_number == self.phone_number