+++ sdk/storage/azure-mgmt-storagecache/generated_samples/aml_filesystems_create_or_update.py
@@ -58,6 +58,12 @@ def main():
                     }
                 },
                 "maintenanceWindow": {"dayOfWeek": "Friday", "timeOfDayUTC": "22:00"},
+                "rootSquashSettings": {
+                    "mode": "All",
+                    "noSquashNidLists": "10.0.0.[5-6]@tcp;10.0.1.2@tcp",
+                    "squashGID": 99,
+                    "squashUID": 99,
+                },
                 "storageCapacityTiB": 16,
             },
             "sku": {"name": "AMLFS-Durable-Premium-250"},
@@ -68,6 +74,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: specification/storagecache/resource-manager/Microsoft.StorageCache/stable/2023-05-01/examples/amlFilesystems_CreateOrUpdate.json
+# x-ms-original-file: specification/storagecache/resource-manager/Microsoft.StorageCache/preview/2023-11-01-preview/examples/amlFilesystems_CreateOrUpdate.json
 if __name__ == "__main__":
     main()